openspiel_requirements_jax.txt
absl-py==2.3.1
attrs==25.3.0
chex==0.1.91
contourpy==1.3.3
cycler==0.12.1
dm-haiku==0.0.15
fonttools==4.60.0
jax==0.7.2
jax-cuda12-pjrt==0.7.2
jax-cuda12-plugin==0.7.2
jaxlib==0.7.2
jmp==0.0.4
kiwisolver==1.4.9
matplotlib==3.10.6
ml_collections==1.1.0
ml_dtypes==0.5.3
numpy==2.3.3
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.13.1.26
nvidia-cufft-cu12==11.4.1.4
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-nccl-cu12==2.28.3
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvshmem-cu12==3.4.5
-e git+https://github.com/google-deepmind/open_spiel.git@33cbb8eb4de26e78b2bb09295900e7883b435e34#egg=open_spiel
opt_einsum==3.4.0
optax==0.2.6
packaging==25.0
pillow==11.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
scipy==1.16.2
six==1.17.0
tabulate==0.9.0
toolz==1.0.0
typing_extensions==4.15.0
